pyproject.toml
[project]
name = "prompt-master"
version = "0.3.0"
description = "AI-powered prompt auditing tool with CLI and API."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"}
]

dependencies = [
    "google-generativeai>=0.8.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0"
]

[project.scripts]
prompt-master = "prompt_master.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_master"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
    "black>=23.0",
    "ruff>=0.1.0"
]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
lint.select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
    "SIM"
]
exclude = [
    ".git",
    "__pycache__",
    "docs",
    "venv",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
    "--cov=src/prompt_master",
    "--cov-report=term-missing",
    "--cov-branch",
]
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for full workflows",
    "slow: Tests that take longer to run",
    "api: Tests for API endpoints",
    "cli: Tests for CLI commands",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/prompt_master"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/virtualenv/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]